boards/dts/soc: syna: sr100: re-write for HWMv2 (define cpuclusters)
Add subfolder 'astra_sr' & introduce cpuclusters. Support for the second
core (M4) of the SR100 will be added in a later commit.
Update copyright statement to use SPDX style.
