+++ boards/syna/astra_sr/sr100/Kconfig.sr100_rdk
@@ -1,7 +1,7 @@
 # Synaptics SR100 Astra Machina Micro (RDK) Board configuration options
 
-# Copyright (c) 2025 Synaptics, Inc.
+# SPDX-FileCopyrightText: Copyright (c) 2025 Synaptics Incorporated
 # SPDX-License-Identifier: Apache-2.0
 
 config BOARD_SR100_RDK
-	select SOC_SR100
+	select SOC_SR100_M55 if BOARD_SR100_RDK_SR100_M55
