+++ boards/syna/astra_sr/sr100/board.cmake
