+++ boards/syna/astra_sr/sr100/support/openocd.cfg
