+++ soc/syna/astra_sr/CMakeLists.txt
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 Synaptics Incorporated
+# SPDX-License-Identifier: Apache-2.0
+
+add_subdirectory(${SOC_SERIES})
