+++ soc/syna/astra_sr/sr1xx/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright (c) 2025 Synaptics Inc.
+# SPDX-FileCopyrightText: Copyright (c) 2025 Synaptics Incorporated
 # SPDX-License-Identifier: Apache-2.0
 
 zephyr_sources(soc.c)
