+++ soc/syna/astra_sr/sr1xx/Kconfig
@@ -1,7 +1,7 @@
-# Copyright (c) 2025 Synaptics, Inc.
+# SPDX-FileCopyrightText: Copyright (c) 2025 Synaptics Incorporated
 # SPDX-License-Identifier: Apache-2.0
 
-config SOC_SR100
+config SOC_SR100_M55
 	select ARM
 	select ARM_MPU
 	select ARMV8_M_DSP
@@ -22,6 +22,3 @@ config SOC_SR100
 	select HAS_CMSIS_CORE
 	select SOC_EARLY_INIT_HOOK
 	select RESET
-
-config SOC_PART_NUMBER
-	default "SR100" if SOC_SR100
