+++ soc/syna/astra_sr/sr1xx/Kconfig.defconfig
@@ -1,4 +1,4 @@
-# Copyright (c) 2025 Synaptics, Inc.
+# SPDX-FileCopyrightText: Copyright (c) 2025 Synaptics Incorporated
 # SPDX-License-Identifier: Apache-2.0
 
 if SOC_SERIES_SR1XX
