+++ soc/syna/astra_sr/sr1xx/soc.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025 Synaptics Incorporated.
+ * SPDX-FileCopyrightText: Copyright (c) 2025 Synaptics Incorporated
  *
  * SPDX-License-Identifier: Apache-2.0
  */
