--- soc/syna/sr100/soc.yml
@@ -1,4 +1,0 @@
-series:
-- name: sr1xx
-  socs:
-  - name: sr100
